OLMIS-4760: made total losses and adjustemnts colum stock based

+++ src/main/java/org/openlmis/requisition/dto/StockCardRangeSummaryDto.java
@@ -35,7 +35,7 @@ public final class StockCardRangeSummaryDto {
 
   @Getter
   @Setter
-  private Long stockOutDays;
+  private Integer stockOutDays;
 
   @Getter
   @Setter
